Refactor chat message sections out of ChatDialogMessages
